pyproject.toml
[tool.poetry]
name = "sadhu"
version = "0.1.0"
description = ""
authors = ["Thanathip Limna <boatkrap@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.14"
Flask = "^3.1.2"
Authlib = "^1.3.2"
Flask-Login = "^0.6.3"
Markdown = "^3.4.1"
Pygments = "^2.13.0"
python-markdown-math = "^0.8"
wheel = "^0.37.1"
flask-wtf = "^1.1.1"
email-validator = "^2.0.0.post2"
requests = "^2.32.3"
flask-mongoengine-3 = "^1.1.0"
brython = "^3.14.0"

[poetry.group.dev.dependencies]
black = "^22.8.0"
watchdog = "^2.1.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
sadhu-web = 'sadhu.cmd.web:main'
sadhu-checker = 'sadhu.cmd.checker:main'


[dependency-groups]
dev = [
    "livereload (>=2.7.1,<3.0.0)",
    "watchdog (>=6.0.0,<7.0.0)"
]
